[refactoring] ICommand.h의 변경을 최소화하기 위해 command header 파일 분리
