+++ SSD/SSD/SSD.vcxproj
@@ -150,7 +150,11 @@
     <None Include="packages.config" />
   </ItemGroup>
   <ItemGroup>
+    <ClInclude Include="command_erase.h" />
+    <ClInclude Include="command_flush.h" />
     <ClInclude Include="command_handler.h" />
+    <ClInclude Include="command_read.h" />
+    <ClInclude Include="command_write.h" />
     <ClInclude Include="nand_handler.h" />
     <ClInclude Include="output_handler.h" />
   </ItemGroup>
