+++ SSD/SSD/command_flush.h
@@ -1,0 +1,14 @@
+#pragma once
+#include "ICommand.h"
+
+class FlushCommand : public ICommand {
+public:
+	FlushCommand(NandHandler* nandHandler);
+	bool isValid(const vector<string>& param) override;
+	void execute(const vector<string>& param) override;
+
+private:
+	void flushBuffer();
+
+	const int PARAMETER_COUNT = 1;
+};
