+++ SSD/SSD/command_handler.h
@@ -7,11 +7,10 @@ public:
 	static std::unique_ptr<ICommand> makeCommand(const string& cmd);
 };
 
-
 class CommandHandler {
 public:
 	CommandHandler();
-    CommandHandler(NandHandler* handler);
+	CommandHandler(NandHandler* handler);
 
 	bool handleCommand(const vector<string>& cmdArr);
 
@@ -19,6 +18,6 @@ private:
 	bool isValidCommand(const vector<string>& cmdArr);
 	void executeCommand(const vector<string>& cmdArr);
 
-    std::unique_ptr<ICommand> m_command;
-    NandHandler* m_nandHandler = nullptr;
+	std::unique_ptr<ICommand> m_command;
+	NandHandler* m_nandHandler = nullptr;
 };
