+++ SSD/SSD/command_write.cpp
@@ -1,4 +1,4 @@
-#include "ICommand.h"
+#include "command_write.h"
 
 WriteCommand::WriteCommand(NandHandler* nandHandler) : ICommand(nandHandler) { }
